Move mode pages
